Revise exchange. Use tag to distinguish send/recv pairs.
When there is only one process in, say, x-direction, a process
will send/receive itself the boundary grid data to realize
the periodic boundary condition. There are two pairs of
the send/receive communication (for two boundaries).
Before this revise, the two pairs were mixed because the
same tag (=0) was used for the two pairs.